Revert calling from_json() in enforce_type() unconditionally.
This change would cause test failures in the LMS, though I can't figure out
why.  This part of the original commit was more of a clean-up than a fix, so
it's not necessary to address the bug with DateTime deserialisation.

I still think the clean-up would actually make sense, but I currently don't have
the time for further debugging.